+++ tests/test_process_engine.py
@@ -4,7 +4,7 @@ from fuseline import Workflow
 from fuseline.broker import MemoryBroker
 from fuseline.broker.clients import LocalBrokerClient
 from fuseline.worker import ProcessEngine
-from fuseline.policies import RetryPolicy
+from fuseline.workflow.policies import RetryPolicy
 from fuseline.workflow import Status, Task
